third commit

+++ app/src/main/java/com/example/weatherapp/modals/Sys.kt
@@ -6,6 +6,6 @@ data class Sys(
 //    val type : Int,
 //    val id : Int,
     val country : String,
-    val sunrise :Int,
-    val Sunset : Int
+    val sunrise :Long,
+    val Sunset : Long
 ) : Serializable
